[JBIDE-9854] created test
git-svn-id: http://svn.jboss.org/repos/jbosstools/trunk@35401 a97e2381-89e5-4abb-bab3-167db6db766c

+++ as/tests/org.jboss.tools.openshift.express.client.test/src/org/jboss/tools/openshift/express/internal/client/test/utils/ApplicationUtils.java
@@ -32,6 +32,9 @@ public class ApplicationUtils {
 	
 	public static void silentlyDestroyAS7Application(String name, User user, IOpenshiftService service) {
 		try {
+			if (name == null) {
+				return;
+			}
 			service.destroyApplication(name, ICartridge.JBOSSAS_7, user);
 		} catch (OpenshiftException e) {
 			e.printStackTrace();
